requirements.txt
google-generativeai
google-adk
google-cloud-aiplatform
google-api-python-client<3.0.0,>=2.157.0
python-dotenv
requests
uvicorn
beautifulsoup4
loguru
pydantic
fastapi
python-multipart
python-pptx>=0.6.21